AP_NavEKF: Prevent flickering bad health status on ground
If the health status is checked after a non-height position or velocity fusion, then the health status will return false.
The height fusion health status has been removed from the check to prevent this.
The height fusion health check cannot be used unless the pre-arm fusion of position and velocity is timed to coincide with the baro fusion.